Convert addon view (base) and log
